additional clang-tidy (mostly boolean logic cleanup)

+++ src/Riff.cpp
@@ -112,7 +112,7 @@ int RiffFinishWriteFile() {
   SetFilePointer(g_hRiffFile, dwDataOffset, NULL, FILE_BEGIN);
   WriteFile(g_hRiffFile, &temp32, 4, &dwNumberOfBytesWritten, NULL);
 
-  return CloseHandle(g_hRiffFile);
+  return static_cast<int>(CloseHandle(g_hRiffFile));
 }
 
 int RiffPutSamples(short *buf, unsigned int uSamples) {
